requirements.txt
flask==2.3.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
werkzeug==3.0.1